+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firBasedSymbol.kt
@@ -1,9 +1,12 @@
 package com.javiersc.kotlin.compiler.extensions.fir
 
+import kotlin.contracts.contract
 import org.jetbrains.kotlin.fir.getOwnerLookupTag
 import org.jetbrains.kotlin.fir.symbols.FirBasedSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirCallableSymbol
 import org.jetbrains.kotlin.fir.symbols.impl.FirClassLikeSymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirFunctionSymbol
+import org.jetbrains.kotlin.fir.symbols.impl.FirPropertySymbol
 import org.jetbrains.kotlin.name.FqName
 
 public fun FirBasedSymbol<*>.packageFqName(): FqName? {
@@ -16,3 +19,13 @@ public fun FirBasedSymbol<*>.packageFqName(): FqName? {
 
 public val FirBasedSymbol<*>.hasOwner: Boolean
     get() = runCatching { getOwnerLookupTag() }.getOrNull() != null
+
+public fun FirBasedSymbol<*>.isProperty(): Boolean {
+    contract { returns(true) implies (this@FirBasedSymbol is FirPropertySymbol) }
+    return this is FirPropertySymbol
+}
+
+public fun FirBasedSymbol<*>.isFunction(): Boolean {
+    contract { returns(true) implies (this@FirBasedSymbol is FirFunctionSymbol) }
+    return this is FirFunctionSymbol
+}